Move CookiesFieldResetButton to its own file
It's been moved back to the Cookies folder
