+++ assets/src/blocks/Cookies/CookiesFieldResetButton.js
@@ -2,7 +2,7 @@ const {__} = wp.i18n;
 const {Tooltip} = wp.components;
 
 export const CookiesFieldResetButton = ({fieldName, toAttribute, currentValue}) => {
-  const COOKIES_DEFAULT_COPY = window.p4_vars.options.cookies_default_copy || {};
+  const COOKIES_DEFAULT_COPY = window.p4_vars?.options?.cookies_default_copy || {};
   const defaultValue = COOKIES_DEFAULT_COPY[fieldName] || '';
 
   if (!currentValue || !defaultValue || currentValue === defaultValue) {
